pyproject.toml
[project]
name = "ios-prep-pipeline"
version = "0.1.0"
description = "RAG-based pipeline to generate iOS interview flashcards"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*"]
